convert ObjectMapper to interface
