+++ configurate-core/src/test/java/ninja/leaping/configurate/objectmapping/ObjectMapperTest.java
@@ -88,7 +88,7 @@ public class ObjectMapperTest {
     @Test
     public void testCommentsApplied() throws ObjectMappingException {
         CommentedConfigurationNode node = SimpleCommentedConfigurationNode.root();
-        ObjectMapper<CommentedObject>.BoundInstance mapper = ObjectMapper.forClass(CommentedObject.class).bindToNew();
+        ObjectMapper.BoundInstance<CommentedObject> mapper = ObjectMapper.forClass(CommentedObject.class).bindToNew();
         CommentedObject obj = mapper.populate(node);
         obj.color = "fuchsia";
         obj.politician = "All of them";
@@ -162,7 +162,7 @@ public class ObjectMapperTest {
     @Test
     public void testNestedObjectWithComments() throws ObjectMappingException {
         CommentedConfigurationNode node = SimpleCommentedConfigurationNode.root();
-        final ObjectMapper<ParentObject>.BoundInstance mapper = ObjectMapper.forObject(new ParentObject());
+        final ObjectMapper.BoundInstance<ParentObject> mapper = ObjectMapper.forObject(new ParentObject());
         mapper.populate(node);
         assertEquals("Comment on parent", node.getNode("inner").getComment().get());
         assertTrue(node.getNode("inner").hasMapChildren());
